Fix and move occ product config providers
