+++ projects/core/src/occ/adapters/product/default-occ-product-config.ts
@@ -1,4 +1,4 @@
-import { OccConfig } from '../../occ/config/occ-config';
+import { OccConfig } from '../../config/occ-config';
 
 export const defaultOccProductConfig: OccConfig = {
   backend: {
