+++ projects/core/src/occ/adapters/product/product-occ.module.ts
@@ -1,8 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { ConfigModule } from '../../../config';
-import { defaultOccProductConfig } from '../../../product/config/product-config';
+import { ConfigModule } from '../../../config/config.module';
 import { PRODUCT_NORMALIZER } from '../../../product/connectors/product/converters';
 import { ProductAdapter } from '../../../product/connectors/product/product.adapter';
 import { PRODUCT_REFERENCES_NORMALIZER } from '../../../product/connectors/references/converters';
@@ -18,6 +17,7 @@ import { OccProductReviewsAdapter } from './occ-product-reviews.adapter';
 import { OccProductSearchAdapter } from './occ-product-search.adapter';
 import { OccProductAdapter } from './occ-product.adapter';
 import { ProductNameNormalizer } from './converters/product-name-normalizer';
+import { defaultOccProductConfig } from './default-occ-product-config';
 
 @NgModule({
   imports: [
