+++ projects/core/src/product/store/effects/product-reviews.effect.spec.ts
@@ -9,7 +9,7 @@ import { cold, hot } from 'jasmine-marbles';
 
 import { Observable, of } from 'rxjs';
 import { OccConfig } from '../../../occ/config/occ-config';
-import { defaultOccProductConfig } from '../../config/product-config';
+import { defaultOccProductConfig } from '../../../occ/adapters/product/product-config';
 import * as fromActions from '../actions/product-reviews.action';
 import * as fromEffects from '../effects/product-reviews.effect';
 import createSpy = jasmine.createSpy;
